Finish the UserRepo class.
UserRepo is a class that connects and operates the User table in the database. With the API of this class, we can insert a user into the table, get user entity by its username, get user's id by its username or a user entity, and check if the username and password is legal.