Fix the bug about the resultSet.
When there is no required data in the database, the resultSet wouldn't be null, but be empry.
Now if there is no required data, the implements will return null or a special value when the return value is a Object or int, but a empty set when the return value is a Iterable Interface.