Finish the ListRepo Class.
This class is responsible for the operations obout common user's shopping list. To make search more efficient, the code will build a independent table for each common user, and it is automatic--every time a instance for a specific user initailized, it will exam if the table for the user has been created, if not, it will create one.